display relative time on notification, and refactor
